app/requirements.txt
fastapi
uvicorn
agno>=2.1.10
httpx
psycopg2-binary
sqlalchemy
ollama
